feat(cli): remove moneyd and standalone spsp commands
the node is the main supported way of interacting with interledger.rs now. in the future we may bring back standalone bundles of functionality if we find a use case for them

BREAKING CHANGES:
- remove spsp pay and server commands from CLI
- remove moneyd command from CLI
