+++ crates/interledger/src/main.rs
@@ -1,30 +1,21 @@
-use base64;
-use clap::{crate_version, App, AppSettings, Arg, ArgGroup, ArgMatches, SubCommand};
+use clap::{crate_version, App, AppSettings, Arg, ArgMatches, SubCommand};
 use config::{Config, Source};
 use config::{FileFormat, Value};
-use futures::future::Future;
+use futures::Future;
 use hex;
-use interledger::{cli::*, node::*};
-use interledger_ildcp::IldcpResponseBuilder;
+use interledger::node::{insert_account_redis, tokio_run, AccountDetails, InterledgerNode};
 use interledger_packet::Address;
 use interledger_service::Username;
 use libc::{c_int, isatty};
 use serde::Deserialize;
 use std::ffi::{OsStr, OsString};
 use std::io::Read;
-use std::net::SocketAddr;
 use std::str::FromStr;
 use std::vec::Vec;
 use url::Url;
 
 pub fn main() {
     env_logger::init();
-
-    let moneyd_uri = format!(
-        "btp+ws://{}:{}@localhost:7768",
-        random_token(),
-        random_token()
-    );
 
     // The naming convention of arguments
     //
@@ -50,118 +41,6 @@ pub fn main() {
         // https://github.com/clap-rs/clap/issues/1536
         .after_help("")
         .subcommands(vec![
-            SubCommand::with_name("spsp")
-                .about("Client and Server for the Simple Payment Setup Protocol (SPSP)")
-                .setting(AppSettings::SubcommandsNegateReqs)
-                .subcommands(vec![
-                    SubCommand::with_name("server")
-                        .about("Run an SPSP Server that automatically accepts incoming money")
-                        .args(&[
-                            // Positional arguments
-                            Arg::with_name("config")
-                                .takes_value(true)
-                                .index(1)
-                                .help("Name of config file (in JSON, YAML, or TOML format)"),
-                            // Non-positional arguments
-                            Arg::with_name("port")
-                                .long("port")
-                                .short("p")
-                                .takes_value(true)
-                                .help("Port that the SPSP server should listen on"),
-                            Arg::with_name("spsp_bind_address")
-                                .long("spsp_bind_address")
-                                .takes_value(true)
-                                .help("The address which this SPSP server is bound to.")
-                                .conflicts_with("port"),
-                            Arg::with_name("btp_server_url")
-                                .long("btp_server_url")
-                                .default_value(&moneyd_uri)
-                                .help("URI of a moneyd or BTP Server to listen on"),
-                            Arg::with_name("ilp_over_http")
-                                .long("use_ilp_over_http")
-                                .help("Accept ILP packets sent over HTTP instead of connecting to a BTP server"),
-                            Arg::with_name("ilp_address")
-                                .long("ilp_address")
-                                .takes_value(true)
-                                .required(true)
-                                .help("The server's ILP address (Required for ilp_over_http)"),
-                            Arg::with_name("incoming_auth_token")
-                                .long("incoming_auth_token")
-                                .takes_value(true)
-                                .required(true)
-                                .help("Token that must be used to authenticate incoming requests (Required for ilp_over_http)"),
-                            Arg::with_name("quiet")
-                                .long("quiet")
-                                .help("Suppress log output"),
-                        ])
-                        .group(ArgGroup::with_name("http_options").requires_all(&["ilp_over_http", "ilp_address", "incoming_auth_token"])),
-                    SubCommand::with_name("pay")
-                        .about("Send an SPSP payment")
-                        .args(&[
-                            // Positional arguments
-                            Arg::with_name("config")
-                                .takes_value(true)
-                                .index(1)
-                                .help("Name of config file (in JSON, YAML, or TOML format)"),
-                            // Non-positional arguments
-                            Arg::with_name("btp_server_url")
-                                .long("btp_server_url")
-                                .takes_value(true)
-                                .help("URI of a moneyd or BTP Server to pay from"),
-                            Arg::with_name("http_server_url")
-                                .long("http_server_url")
-                                .takes_value(true)
-                                .help("HTTP URL of the connector to pay from"),
-                            Arg::with_name("receiver")
-                                .long("receiver")
-                                .short("r")
-                                .takes_value(true)
-                                .required(true)
-                                .help("Payment Pointer of the receiver"),
-                            Arg::with_name("amount")
-                                .long("amount")
-                                .short("a")
-                                .takes_value(true)
-                                .required(true)
-                                .help("Amount to send, denominated in the connector's units"),
-                            Arg::with_name("quiet")
-                                .long("quiet")
-                                .help("Suppress log output"),
-                        ]),
-                ]),
-            SubCommand::with_name("moneyd")
-                .about("Run a local connector that exposes a BTP server with open signup")
-                .setting(AppSettings::SubcommandsNegateReqs)
-                .subcommand(SubCommand::with_name("local")
-                    .about("Run locally without connecting to a remote connector")
-                    .args(&[
-                        // Positional arguments
-                        Arg::with_name("config")
-                            .takes_value(true)
-                            .index(1)
-                            .help("Name of config file (in JSON, YAML, or TOML format)"),
-                        // Non-positional arguments
-                        Arg::with_name("port")
-                            .long("port")
-                            .short("p")
-                            .help("Port to listen for BTP connections on"),
-                        Arg::with_name("btp_bind_address")
-                            .long("btp_bind_address")
-                            .takes_value(true)
-                            .help("The address which this moneyd server is bound to.")
-                            .conflicts_with("port"),
-                        Arg::with_name("ilp_address")
-                            .long("ilp_address")
-                            .default_value("private.local"),
-                        Arg::with_name("asset_code")
-                            .long("asset_code")
-                            .default_value("XYZ"),
-                        Arg::with_name("asset_scale")
-                            .long("asset_scale")
-                            .help("Scale of the asset this account's balance is denominated in (a scale of 2 means that 100.50 will be represented as 10050) Refer to https://bit.ly/2ZlOy9n")
-                            .default_value("9"),
-                    ])
-                ),
             SubCommand::with_name("node")
                 .about("Run an Interledger node (sender, connector, receiver bundle)")
                 .setting(AppSettings::SubcommandsNegateReqs)
@@ -338,24 +217,6 @@ pub fn main() {
     }
     let matches = app.clone().get_matches();
     match matches.subcommand() {
-        ("spsp", Some(spsp_matches)) => match spsp_matches.subcommand() {
-            ("server", Some(spsp_server_matches)) => {
-                merge_args(&mut config, &spsp_server_matches);
-                run_spsp_server(config.try_into::<SpspServerOpt>().unwrap());
-            }
-            ("pay", Some(spsp_pay_matches)) => {
-                merge_args(&mut config, &spsp_pay_matches);
-                run_spsp_pay(config.try_into::<SpspPayOpt>().unwrap());
-            }
-            _ => println!("{}", spsp_matches.usage()),
-        },
-        ("moneyd", Some(moneyd_matches)) => match moneyd_matches.subcommand() {
-            ("local", Some(moneyd_local_matches)) => {
-                merge_args(&mut config, &moneyd_local_matches);
-                run_moneyd_local(config.try_into::<MoneydLocalOpt>().unwrap());
-            }
-            _ => println!("{}", moneyd_matches.usage()),
-        },
         ("node", Some(node_matches)) => match node_matches.subcommand() {
             ("accounts", Some(node_accounts_matches)) => match node_accounts_matches.subcommand() {
                 ("add", Some(node_accounts_add_matches)) => {
@@ -515,63 +376,6 @@ fn is_fd_tty(file_descriptor: c_int) -> bool {
     result == 1
 }
 
-fn run_spsp_server(opt: SpspServerOpt) {
-    if opt.ilp_over_http {
-        let ildcp_info = IldcpResponseBuilder {
-            client_address: &Address::from_str(&opt.ilp_address).unwrap(),
-            asset_code: "",
-            asset_scale: 0,
-        }
-        .build();
-        tokio_run(run_spsp_server_http(
-            ildcp_info,
-            opt.get_bind_address(),
-            opt.incoming_auth_token.clone(),
-            opt.quiet,
-        ));
-    } else {
-        tokio_run(run_spsp_server_btp(
-            &opt.btp_server_url,
-            opt.get_bind_address(),
-            opt.quiet,
-        ));
-    }
-}
-
-fn run_spsp_pay(opt: SpspPayOpt) {
-    // Check for http_server first because btp_server has the default value of connecting to moneyd
-    if let Some(http_server_url) = &opt.http_server_url {
-        tokio_run(send_spsp_payment_http(
-            http_server_url,
-            &opt.receiver,
-            opt.amount,
-            opt.quiet,
-        ));
-    } else if let Some(btp_server_url) = &opt.btp_server_url {
-        tokio_run(send_spsp_payment_btp(
-            btp_server_url,
-            &opt.receiver,
-            opt.amount,
-            opt.quiet,
-        ));
-    } else {
-        panic!("Must specify either btp_server or http_server");
-    }
-}
-
-fn run_moneyd_local(opt: MoneydLocalOpt) {
-    let ildcp_info = IldcpResponseBuilder {
-        client_address: &Address::from_str(&opt.ilp_address).unwrap(),
-        asset_code: &opt.asset_code,
-        asset_scale: opt.asset_scale,
-    }
-    .build();
-    tokio_run(interledger::cli::run_moneyd_local(
-        opt.get_bind_address(),
-        ildcp_info,
-    ));
-}
-
 fn run_node_accounts_add(opt: NodeAccountsAddOpt) {
     let (http_endpoint, http_outgoing_token) = if let Some(url) = &opt.http_server_url {
         let url = Url::parse(url).expect("Invalid URL");
@@ -625,62 +429,6 @@ fn run_node_accounts_add(opt: NodeAccountsAddOpt) {
         settlement_engine_url: None,
     };
     tokio_run(insert_account_redis(redis_url, &server_secret, account).and_then(move |_| Ok(())));
-}
-
-#[derive(Deserialize, Clone)]
-struct SpspServerOpt {
-    port: Option<u16>,
-    spsp_bind_address: Option<SocketAddr>,
-    btp_server_url: String,
-    #[serde(default)]
-    ilp_over_http: bool,
-    ilp_address: String,
-    incoming_auth_token: String,
-    #[serde(default)]
-    quiet: bool,
-}
-
-impl SpspServerOpt {
-    fn get_bind_address(&self) -> SocketAddr {
-        if let Some(address) = self.spsp_bind_address {
-            return address;
-        }
-        match self.port {
-            Some(port) => SocketAddr::from(([127, 0, 0, 1], port)),
-            None => SocketAddr::from(([127, 0, 0, 1], 3000)),
-        }
-    }
-}
-
-#[derive(Deserialize, Clone)]
-struct SpspPayOpt {
-    btp_server_url: Option<String>,
-    http_server_url: Option<String>,
-    receiver: String,
-    amount: u64,
-    #[serde(default)]
-    quiet: bool,
-}
-
-#[derive(Deserialize, Clone)]
-struct MoneydLocalOpt {
-    port: Option<u16>,
-    btp_bind_address: Option<SocketAddr>,
-    ilp_address: String,
-    asset_code: String,
-    asset_scale: u8,
-}
-
-impl MoneydLocalOpt {
-    fn get_bind_address(&self) -> SocketAddr {
-        if let Some(address) = self.btp_bind_address {
-            return address;
-        }
-        match self.port {
-            Some(port) => SocketAddr::from(([127, 0, 0, 1], port)),
-            None => SocketAddr::from(([127, 0, 0, 1], 7768)),
-        }
-    }
 }
 
 #[derive(Deserialize, Clone)]
